WIP (SERVER-339) Remove all post install pkg actions
Without this patch the commands executed post package installation are
incorrect and operate against paths that are no longer relevant with the
AIO layout.  This is a problem because the puppet server service fails
to start immediately after package installation.  There are surely many
more issues, but the most obvious is that things simply don't work.

The post install commands are being removed entirely because the state
of the system should be specified entirely by the package manifest, not
by a script executed post package installation.  It's unclear which
install actions are necessary or not, so all of them are being removed
to see if the new AIO puppet-agent package has correctly specified all
of the things puppet-server depends upon.