(SERVER-339) Override beaker default AIO paths for centos7
Without this patch the paths assumed by beaker are incorrect.  This
patch addresses the problem by overriding the paths explicitly in the
host configuration.

This is not ideal, ideally beaker would have no knowledge of platform
specific details, however it does and we cannot wait around for beaker
to be released so this patch resolves the problem by explicitly
overriding the paths for centos7.

The other host configurations should be updated once acceptance tests
are green on CentOS 7.