(SERVER-339) Allow for beaker from HEAD of master
Without this patch the version of beaker is hard-coded in the Gemfile.
This is a problem because the version of beaker that is hard-coded is
incompatible with new AIO puppet-agent paths.  This patch addresses the
problem by allowing bundler to install beaker from the HEAD of the
master branch in git.  This is a less than ideal, however it is fast and
we need not wait for the full release cycle of beaker in order to get
our tests passing.

In order to use beaker from master:

    export BEAKER_LOCATION='git://github.com/beaker#master'
    bundle install --path .bundle/gems/
    ...
    Using beaker 2.4.1 (was 2.3.0) from git://github.com/puppetlabs/beaker (at master)