(SERVER-548) Fail the unit tests if header assertions are skipped
Without this patch there is the possibility that the assertions made
against the munged request are not executed.  This is a problem because
there could be an issue with the routes which are not caught by the unit
tests.  This patch addresses the problem by creating an Atom and
wrapping the assertions made from a downstream middleware in a function
that updates the Atom when triggered.  The patch makes a subsequent
assertion against the Atom to make sure it was updated by the
assertion function.