(SERVER-297) Add jruby subcommand integration tests
Without this patch there are not tests covering the behavior of the irb,
ruby, and gem JRuby subcommands.  This patch addresses the problem by
adding integration tests that assert against standard output and the
exit status.

NB this patch differs from the patch against master in that the
ruby-load-path setting is under the os-settings configuration section in
this patch while it is under the jruby-puppet section in the patch
against master for puppetserver 2.0