(SERVER-769) Add FEATURE_TK_AUTHZ=true flag
Without this patch puppetserver performs no authorization for the Puppet
master service.  We want to allow authorization to be configurable
through the use of a feature flag because we want to either handle
authorization checking in puppetserver or handle it inside of Puppet.

This patch addresses the problem by implementing a very simple check
against the FEATURE_TK_AUTHZ environment variable.  If the variable is
"true" then the master service ring handler is wrapped with a middleware
that performs authorization using trapperkeeper-authorization.
Otherwise, no wrapping occurs and Puppet's internal authconfig behavior
is used.

Puppet is configured to enable or disable authconfig authorization
depending upon the FEATURE_TK_AUTHZ flag.