(SERVER-769) Ensure all services are wrapped with tk-authz
Without this patch only the requests which previously used a certificate
whitelist and the puppet master service are wrapped in
trapperkeeper-authorization checks.

Instead of this behavior, we want all requests to be wrapped in tk-authz
checks if a feature flag is set.  We also want the feature flag to be
able to be turned off which would cause only a subset of requests to
flip over from being handled by tk-authz to Puppet's authconfig.

The following requests will be handled by trapperkeeper-authorization
instead of Puppet authconfig if `FEATURE_TK_AUTHZ=true`.

 * master-service on /puppet (✓)
 * legacy-routes-service on "" (✓)

All of these other requests will always be handled by
trapperkeeper-authorization:

 * certificate-authority-service on "/puppet-ca" (✓)
 * puppet-admin-service on "/puppet-admin-api" (✓)

In all cases the authorization check is handled by a ring middleware
close to the handler.  This usually happens in the wrap-middleware
function implemented in the service's core namespace.

NOTE: This patch does not update the internal rules listing.  As a
result, a custom rules.conf file may need to be laid down to get this to
work.  A follow on patch should update the internal list of
authorization rules.