(SERVER-769) Fix bug where bypass_authorization is always true
Without this patch bypass_authorization is always true.  This is a
problem because it should be set to false when FEATURE_TK_AUTHZ is not
set to a value of "true".

The root cause of the problem is that I assumed puppet configuration is
passed in a way it is not actually passed.  Puppet Server passes
configuration values through the scripting container argument vector,
which means that boolean settings must be passed as either --foo to
enable and --no-foo to disable.

This is contrary to the configuration file where foo=true and foo=false
to set the corresponding values.

This patch addresses the problem by passing the options as
--bypass_authorization or --no-bypass_authorization depending on the
value of FEATURE_TK_AUTHZ.